Update and fix the `rake changelog` task
Now that the project was renamed, we can use the final name for the
project.  While here, fix style and update the copy-pasta to correspond
to the project.

Also remove obsolete development dependency that prevent GCG from
starting, and use the name of the *last* tag, not the *next* tag to
fill-in the blanks of the history file.