Lower the version of the gem
The Perforce / PuppetLabs tooling bumped the version of gems to the next
minor version when releasing a version, while at Vox Pupuli we usually
bump the patch version and add a `.rc1` to make sure version never go
backwards.

Adopt this scheme by lowering the gem version before we release it for
the first time.

4.11.0.rc1 is before 4.11.0.