pin rubocop to < 1.79
The whole rubocop setup is weird. Facter says it's compatible with ruby
2.5, but we pull in rubocop >= 1.28 < 2. Probably someone wanted to pin
it to 1.28.0 but forgot the last 0. This PR prevents us from pulling in
new minor versions, which potentially introduce new config options and
break our CI (like the 1.78 release, because of that I'm also
regenerating the .rubocop_todo.yml).

In the future we should investigate the desired lowest Ruby version, I
don't see a reason for pinning to Ruby 2.5. openvoxserver 7 uses jruby
9.3 which is Ruby 2.6.