Rework facter man page
This repository contain a facter(8) man page.  This man page is
generated by running `facter --man` which produce markdown and is passed
to `ronn`, a ruby gem that was last updated in 2010 and bring a bunch of
outdated dependencies that cause trouble with latest Ruby.

In order to get rid of this legacy gem, convert the "man" template so
that it output a man page directly (with proper markup), that way no
conversion is needed and the page is slightly improved (it is still not
the quality of hand crafted man pages, but I do not consider this an
issue given the point where we start).  While here, change the man page
section as facter is not a system manager command, but rather a general
command.  Get rid of the legacy man page and generate a new one.